+++ src/click/model/ipv4-l3-click-protocol.cc
@@ -454,10 +454,7 @@ Ipv4L3ClickProtocol::SourceAddressSelection(uint32_t interfaceIdx, Ipv4Address d
         Ipv4InterfaceAddress test = GetAddress(interfaceIdx, i);
         if (test.GetLocal().CombineMask(test.GetMask()) == dest.CombineMask(test.GetMask()))
         {
-            if (!test.IsSecondary())
-            {
-                return test.GetLocal();
-            }
+            return test.GetLocal();
         }
     }
     return candidate;
@@ -480,10 +477,6 @@ Ipv4L3ClickProtocol::SelectSourceAddress(Ptr<const NetDevice> device,
         for (uint32_t j = 0; j < GetNAddresses(i); j++)
         {
             iaddr = GetAddress(i, j);
-            if (iaddr.IsSecondary())
-            {
-                continue;
-            }
             if (iaddr.GetScope() > scope)
             {
                 continue;
@@ -510,10 +503,6 @@ Ipv4L3ClickProtocol::SelectSourceAddress(Ptr<const NetDevice> device,
         for (uint32_t j = 0; j < GetNAddresses(i); j++)
         {
             iaddr = GetAddress(i, j);
-            if (iaddr.IsSecondary())
-            {
-                continue;
-            }
             if (iaddr.GetScope() != Ipv4InterfaceAddress::LINK && iaddr.GetScope() <= scope)
             {
                 return iaddr.GetLocal();